Better and more flexible auth handling

+++ interfaces/rest.js
@@ -1,6 +1,8 @@
 var restify = require('restify');
 var unknownMethodHandler = require('../utls.js').unknownMethodHandler;
 var authenticate = require('../auth.js').authenticate;
+var hasPermission = require('../auth.js').hasPermission;
+
 var config = require('../config.json');
 var plugins = require("../services/plugins.js").plugins;
 var saveconfig = require('../utls.js').saveconfig;
@@ -10,7 +12,7 @@ var restserver = restify.createServer();
 restserver.use(restify.bodyParser());
 restserver.use(restify.authorizationParser());
 restserver.on('MethodNotAllowed', unknownMethodHandler);
-restserver.use(authenticate);
+//restserver.use(authenticate);
 
 restserver.use(
   function crossOrigin(req,res,next){
@@ -30,8 +32,31 @@ restserver.get('/', function info(req, res, next){
   res.send(response);
 });
 
+function restauth(req, service, permission){
+  if (!('X-Access-Token' in req.headers) && 'x-access-token' in req.headers){
+    req.headers['X-Access-Token'] = req.headers['x-access-token']
+  }
+    
+  if (!('X-Access-Token' in req.headers)){
+    return false;
+  }
+
+  if (!hasPermission(permission, req.headers['X-Access-Token'], service)){
+    return false;
+  }else{
+    return true;
+  }
+}
+
+function unauthorized(res){
+    console.log('not authenticated, missing header');
+    res.writeHead(403);
+    res.end('Sorry you are not authorized');
+    return res
+}
 
 restserver.get('/gameservers/', function info(req, res, next){
+  if (!restauth(req, -1, "services:list")){res = unauthorized(res); return next();}
   response = [];
   servers.forEach(
     function(service){
@@ -41,9 +66,9 @@ restserver.get('/gameservers/', function info(req, res, next){
 });
 
 restserver.post('/gameservers/', function info(req, res, next){
+  if (!restauth(req, -1, "services:new")){res = unauthorized(res); return next();}  
   id = config.servers.push(JSON.parse(req.params['settings']));
-
-  // Stupid java indexes
+  // As push returns the array length, not the id
   id = id - 1;
   saveconfig(config); 
   initServer(id);
@@ -53,6 +78,7 @@ restserver.post('/gameservers/', function info(req, res, next){
 });
 
 restserver.del('/gameservers/:id', function info(req, res, next){
+  if (!restauth(req, req.params.id, "service:delete")){res = unauthorized(res); return next();}  
   service = servers[req.params.id];
   // TODO: if on, turn off
   service.delete();    var ftpd = require('ftpd');
@@ -63,15 +89,15 @@ restserver.del('/gameservers/:id', function info(req, res, next){
 });
 
 restserver.get('/gameservers/:id', function (req, res, next){
-
+  if (!restauth(req, req.params.id, "service:get")){res = unauthorized(res); return next();}
   service = servers[req.params.id];
   res.send(service.info());
 });
 
 
 restserver.put('/gameservers/:id', function info(req, res, next){
+  if (!restauth(req, req.params.id, "service:update")){res = unauthorized(res); return next();}
   service = servers[req.params.id];
-
   config.servers[req.params.id].variables = req.params['variables'];
   saveconfig(config);
   
@@ -79,19 +105,64 @@ restserver.put('/gameservers/:id', function info(req, res, next){
 });
 
 
-restserver.get('/gameservers/:id/on', function on(req, res, next){service = servers[req.params.id]; service.turnon();res.send('ok')});
-restserver.get('/gameservers/:id/off', function off(req, res, next){service = servers[req.params.id]; service.turnoff();res.send('ok')});
-restserver.get('/gameservers/:id/restart', function restart(req, res, next){service = servers[req.params.id]; service.restart();res.send('ok')});
-restserver.get('/gameservers/:id/configlist', function configlist(req, res, next){service = servers[req.params.id]; res.send(service.configlist());});
-restserver.get('/gameservers/:id/maplist', function maplist(req, res, next){service = servers[req.params.id]; res.send(service.maplist());});
-restserver.get('/gameservers/:id/query', function query(req, res, next){service = servers[req.params.id]; res.send(service.lastquery());});
+restserver.get('/gameservers/:id/on', function on(req, res, next){
+  if (!restauth(req, req.params.id, "service:power")){res = unauthorized(res); return next();}
+  service = servers[req.params.id];
+  service.turnon();
+  res.send('ok')
+  
+});
 
-restserver.post('/gameservers/:id/console', function command(req, res, next){service = servers[req.params.id];  res.send(service.send(req.params.command))});
+restserver.get('/gameservers/:id/off', function off(req, res, next){
+  if (!restauth(req, req.params.id, "service:power")){res = unauthorized(res); return next();}
+  service = servers[req.params.id]; 
+  service.turnoff();
+  res.send('ok')
+});
 
-restserver.get('/gameservers/:id/addonsinstalled', function command(req, res, next){service = servers[req.params.id]; res.send(service.addonlist());});
+restserver.get('/gameservers/:id/restart', function restart(req, res, next){
+  if (!restauth(req, req.params.id, "service:power")){res = unauthorized(res); return next();}
+  service = servers[req.params.id];
+  service.restart();
+  res.send('ok')
+});
+restserver.get('/gameservers/:id/configlist', function configlist(req, res, next){
+  if (!restauth(req, req.params.id, "service:file")){res = unauthorized(res); return next();}
+  service = servers[req.params.id]; 
+  res.send(service.configlist());
+});
+restserver.get('/gameservers/:id/maplist', function maplist(req, res, next){
+  if (!restauth(req, req.params.id, "service:file")){res = unauthorized(res); return next();}
+  service = servers[req.params.id]; 
+  res.send(service.maplist());
+});
+restserver.get('/gameservers/:id/query', function query(req, res, next){
+  if (!restauth(req, req.params.id, "service:query")){res = unauthorized(res); return next();}
+  service = servers[req.params.id]; res.send(service.lastquery());
+  
+});
 
-restserver.get(/^\/gameservers\/(\d+)\/file\/(.+)/, function(req, res, next) {service = servers[req.params[0]];res.send({'contents':service.readfile(req.params[1])});});
+restserver.post('/gameservers/:id/console', function command(req, res, next){
+  if (!restauth(req, req.params.id, "service:console")){res = unauthorized(res); return next();}
+  service = servers[req.params.id];
+  res.send(service.send(req.params.command))
+  
+});
+
+restserver.get('/gameservers/:id/addonsinstalled', function command(req, res, next){
+  if (!restauth(req, req.params.id, "service:addons")){res = unauthorized(res); return next();}
+  service = servers[req.params.id];
+  res.send(service.addonlist());
+});
+
+restserver.get(/^\/gameservers\/(\d+)\/file\/(.+)/, function(req, res, next) {
+    if (!restauth(req, req.params.id, "service:file")){res = unauthorized(res); return next();}
+  service = servers[req.params[0]];
+  res.send({'contents':service.readfile(req.params[1])});
+});
+
 restserver.put(/^\/gameservers\/(\d+)\/file\/(.+)/, function(req, res, next) {
+  if (!restauth(req, req.params.id, "service:file")){res = unauthorized(res); return next();}
   if ('contents' in req.params){
     service = servers[req.params[0]];res.send(service.writefile(req.params[1], req.params['contents']));
   }
@@ -99,11 +170,24 @@ restserver.put(/^\/gameservers\/(\d+)\/file\/(.+)/, function(req, res, next) {
     service = servers[req.params[0]];res.send(service.downloadfile(req.params['url'], req.params[1]));
   }
 });
-// TODO : put send
 
-restserver.get('/gameservers/:id/gamemodes', function command(req, res, next){service = servers[req.params.id]; service.getgamemodes(res);});
-restserver.put('/gameservers/:id/gamemodes', function command(req, res, next){service = servers[req.params.id]; service.installgamemode(req.params['gamemode']); res.send("ok");});
-restserver.del('/gameservers/:id/gamemodes', function command(req, res, next){service = servers[req.params.id]; service.getgamemode(res);});
+restserver.get('/gameservers/:id/gamemodes', function command(req, res, next){
+  if (!restauth(req, req.params.id, "gamemodes:get")){res = unauthorized(res); return next();}
+  service = servers[req.params.id];
+  service.getgamemodes(res);
+});
+
+restserver.put('/gameservers/:id/gamemodes', function command(req, res, next){
+  if (!restauth(req, req.params.id, "gamemodes:edit")){res = unauthorized(res); return next();}
+  service = servers[req.params.id];
+  service.installgamemode(req.params['gamemode']);
+  res.send("ok");
+});
+restserver.del('/gameservers/:id/gamemodes', function command(req, res, next){
+  if (!restauth(req, req.params.id, "gamemodes:edit")){res = unauthorized(res); return next();}  
+  service = servers[req.params.id];
+  service.getgamemode(res);
+});
 
 
 restserver.listen(config.daemon.listenport, function() {
